改變介面，分數是double，且增加相關的evaluator

+++ src/main/java/evaluator/EvaluateFunction.java
@@ -2,7 +2,7 @@ package evaluator;
 
 public interface EvaluateFunction<T> {
 
-    int evaluate(T item);
+    double evaluate(T item);
 
     int maxScore(T item);
 }
