+++ src/main/java/evaluator/IntArrayCountEvaluator.java
@@ -3,7 +3,7 @@ package evaluator;
 public class IntArrayCountEvaluator implements EvaluateFunction<int[]> {
 
     @Override
-    public int evaluate(int[] item) {
+    public double evaluate(int[] item) {
         int count = 0;
         for (int element : item) {
             count += element;
