+++ src/main/java/evaluator/IntStreamCountEvaluator.java
@@ -4,7 +4,7 @@ import java.util.stream.IntStream;
 
 public class IntStreamCountEvaluator implements EvaluateFunction<IntStream> {
     @Override
-    public int evaluate(IntStream item) {
+    public double evaluate(IntStream item) {
         return (int) item.filter((x) -> x == 1).count();
     }
